Add support for custom lister on Maven repositories
So far providing a custom lister was only available on Ivy repositories.
With this change, the capability is now migrated to our internal abstract
repository, which makes it possible to use it in both Maven and Ivy.
It's worth noting that only listing is wired now, and it will be possible
to use component metadata suppliers too with Maven repository in a
subsequent commit.