Make snapshotting of Gradle modules faster
Jar files that are part of the current Gradle installation
never change and only depend on the current Gradle version.

This change ensures that they are only snaphshotted once
and avoids actually hitting the file system, but instead
simply computes the snapshot based on the file name and
Gradle version.

This dramatically speeds up test execution for our own
tests.